old/requirements.txt
numpy==2.4.2
pymavlink==2.4.49
pynput==1.8.1
pyrealsense2==2.56.5.9235
pyserial==3.5
Opencv-python>=4.10.0
evdev-binary==1.9.3
numpy==2.4.2
pymavlink==2.4.49
pynput==1.8.1
pyrealsense2==2.56.5.9235
pyserial==3.5
Opencv-python>=4.10.0
PyYAML==6.0.1
